refactor(routes): split users.rs by concern — my_balance + invite move out
Pure reorganization of the 1105-LoC crates/spinbike-server/src/routes/users.rs
into three concern-scoped files. Zero URL / behavior / JSON-shape change.

- NEW routes/my_balance.rs: customer self-service `my_balance` handler +
  `BalanceResponse` / `RecentTx` (GET /api/my/balance). Registered via
  `pub mod my_balance;` + `.merge(my_balance::routes())` in routes/mod.rs.
- EXTEND routes/auth.rs: `invite_user` + `invite_email` + `InviteResponse`
  moved here beside the identical magic-link machinery (request_login_link /
  token_login already redeem PURPOSE_INVITE). POST /api/users/{id}/invite now
  registered in auth::routes(). `db::` → `users::` (same crate::db::users fn).
- routes/users.rs keeps staff-CRUD only (list/search/create/lookup/topup/
  block/negative-balance/update/by-last-movement/delete + user_transactions +
  user_stats), drops the now-unused login_tokens + MailError imports.

Every moved handler keeps its exact extractor (my_balance: AuthUser;
invite_user: StaffUser), URL, method, status codes and response shape. No
test references the routes module path (all drive by HTTP URL), so existing
coverage stays green unchanged.

Closes #165
