+++ crates/spinbike-server/src/routes/mod.rs
@@ -2,6 +2,7 @@ pub mod admin;
 pub mod auth;
 pub mod classes;
 pub mod door;
+pub mod my_balance;
 pub mod payments;
 pub mod persistent_bookings;
 pub mod reports;
@@ -42,6 +43,7 @@ pub fn api_routes() -> Router<AppState> {
         .merge(classes::routes())
         .merge(door::routes())
         .merge(users::routes())
+        .merge(my_balance::routes())
         .merge(payments::routes())
         .merge(admin::routes())
         .merge(persistent_bookings::routes())
